backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pyjwt
bcrypt
motor
python-dotenv
requests
